Create Automatic_DJ_v1.py
This is a proof of concept python file. The important lines here are lines 41-52 and lines 143-157. These are the lines that update the counter in this instance of the swapping test class. By using the .set_input_index() method we are able to create a program that automatically swaps between a list of songs.