Automatic_DJ_V3.py
Latest version of the automatic DJ. This version reduces the used code by 23 lines of code. 
Rather than use a previously defined class that comes generated from a .grc file this code instead creates the class attributes on the fly using the dostuff() function. 
Keep in the mind the do stuff function requires two (2) arguments. The first one must be tb and the second one is your desired transmit sample rate.

There are two functions that have been added into the code as well. The dostuff() function which runs the class creation / block connection, disconnection and reconfiguration, and the song_info() function. The song_info() function returns two variables: the song's sample rate and it's duration in seconds in float format.