Create PTTv6.py
Proof of concept. After a few weeks of effort we have a PTT radio. However it is also a Push to Receive (PTR) radio as well. To transmit the user must press 't' to receive the user must press 'r'. 
This relies upon Tkinter and a GUI. While not ideal, the function is executed. With some further refinement this will allow the hackrf to become an actual walkie talkie.